Fix eth call config
